proposal-generator/requirements.txt
flask==3.1.0
python-dotenv==1.0.1
markdown==3.7
jinja2==3.1.5
weasyprint==63.1
